+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Minimum PHP version is now 7.4
 
+✨ **New ISO currencies**
+
+- `SLE` (Leone) in Sierra Leone (`SL`)
+
 ## [0.5.2](https://github.com/brick/money/releases/tag/0.5.2) - 2021-04-03
 
 ✨ **New methods**