+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 
 - `SLE` (Leone) in Sierra Leone (`SL`)
 
+👌 **Improvements**
+
+- Compatibility with `brick/math` version `0.10`
+
 ## [0.5.2](https://github.com/brick/money/releases/tag/0.5.2) - 2021-04-03
 
 ✨ **New methods**